requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
google-auth==2.25.0
email-validator==2.1.0
python-multipart==0.0.6
requests==2.31.0